Uh, everything's going here, i guess...
I messed up of the commiting or whatever and all my code is in this branch. That's fine; right now i want to fix whatever's wrong with Playfield, merge with master, and continue like nothing happened.

Playfield currently seems to draw the inner and outer circle on a different rotation, at least as far as colloring the segments is concerned.