Gunman falls and points.
That's all it does, yet. There's no gun in its hand so i suppose it's either going to shoot out of its finger, or that'll be handled later.

There's no Bullet class yet, so not shooting is possible.

Collision has not been worked on, so no interacting with Playfield is possible.