Collision between gunmen and playfields works.
Not just that, the normal and gravitational forces combine to create tangental motion! This needs a friction, and i'm worried about how currently the position of the gunman is not really clamped to the playfield. Velocity is redirected, but the position isn't changed upon collision. This may eventually have to change, but it does not cause problems right now.

Gunman must now hold a Playfield reference, always. This allows it to create the normal from whatever position it is at to the Playfield.

After Bullets are added, types will become very relevant to collision. Currently, only collision(playfield,gunman) is handled; i will have to use the double-dispatch solution used in other projects to deal (properly) with bullet-playfield collisions as well as gunman-playfield collisions.