Well, at least the gunman can fire now.
That's the good news, but also, to avoid making actors, the ActorList, globally defined in Actor.h (which leads to not just design problems but technical ones), Actor<T,N>::inserter was added to insert ActorPointers into whatever local ActorList the client wants. But, this meant making the typedefs for ActorPointer and ActorList in Actor.h, which i don't see a big problem with.

This is a real hacky solution and it's great to have come up with a quick-n-dirty one right now, but it will eventually have to be improved. The comment explaining Actor::inserter practically represents that by being shitty itself!

But maybe the idea of having a static inserter function isn't itself bad... the problem is that it contains knowledge of Actor::value_type and N_DIMENSIONS which defeats the purpose of templating Actor.

Perhaps what i really should do is untemplate Actor.